Phase 2: extraction
Adds the extract stage of the pipeline plus the queue + API surface
that drives it. Submitting a URL now enqueues a job, the worker picks
it up, runs Firecrawl extraction with retries, and the status endpoint
reports stage transitions in real time.

Surface
- POST /api/v1/submit (201/400/409): Pydantic SubmitRequest with
  extra=forbid and a custom validator that runs AnyHttpUrl validation
  but preserves the user's exact URL string so episode_id and the
  stored url stay byte-identical to what was sent.
- GET /api/v1/status/{job_id} (200/404): full envelope including
  stage and error.
- Global error envelope {error, status, details?} via FastAPI handlers
  for HTTPException, RequestValidationError (with jsonable_encoder
  guard), and a catch-all Exception that never leaks internals on 500.

Pipeline
- services/extraction.py: async Firecrawl client. POST /v1/scrape with
  tenacity AsyncRetrying (FIRECRAWL_RETRY_COUNT, exponential backoff).
  Typed exceptions: ExtractionTransientError (5xx/network/timeout),
  ExtractionPermanentError (4xx/non-dict body/success=false),
  ExtractionTooShortError (< MIN_EXTRACTION_CHARS). Safe against
  data=null and non-dict JSON bodies.
- services/jobs.py: DB helpers. create_job runs duplicate-check +
  optional reprocess wipe + insert in a single BEGIN IMMEDIATE so
  racing submits cannot both create a job. claim_next_queued is
  atomic. Every UPDATE explicitly bumps updated_at per the
  application-managed timestamps contract.
- services/pipeline.py: process_job(job, settings). Whole job wrapped
  in asyncio.wait_for(JOB_TIMEOUT_SECONDS). Stage writes its name to
  jobs.stage before doing any work, then logs stage_start / stage_end
  / stage_failed (BaseException-catching so timeout-cancelled stages
  still emit telemetry). _finalize_failure is shared between the
  timeout and exception paths and guards both _last_stage and
  _persist_failure so secondary DB errors don't mask the original.
- services/reachability.py: probes Firecrawl by trying /v1/health,
  /health, then / and accepting the first 2xx. Worker calls it at
  startup and sys.exit(1)s on failure; FastAPI lifespan deliberately
  does NOT, so /health/ready stays reachable for triage.
- worker.py: polling loop now actually picks up queued jobs via
  claim_next_queued, calls pipeline.process_job, and wraps the body
  in try/except so a transient DB failure logs and backs off instead
  of killing the worker.

Config / deps
- pyproject: added httpx and tenacity to runtime deps (httpx moved
  out of dev). JOB_TIMEOUT_SECONDS is now float so test overrides
  work.
- .env.example: FIRECRAWL_URL default flipped to
  http://host.docker.internal:3002 so a fresh clone resolves without
  joining an external compose network.

Tests (43 new, 59 total)
- test_extraction.py: happy, 5xx-then-success retry, no-retry on 4xx,
  retry exhaustion, MIN_EXTRACTION_CHARS guard, success=false.
- test_pipeline.py: stage transitions on success, failed with stage
  and error on extraction failure, JOB_TIMEOUT_SECONDS path reports
  the last persisted stage.
- test_api_v1.py: 201/400/409/200/404 plus extra=forbid rejection,
  raw URL preserved through to status, reprocess+inflight 409,
  failed-job status, 500 envelope without leakage.
- test_reachability.py: 2xx OK, fallback through endpoint candidates,
  network unreachable, persistent 5xx reports last detail, run_all
  raises on failure and returns on success.
- test_worker.py: end-to-end pickup runs the pipeline.

Container smoke verified with a mock Firecrawl: reachability picks
/v1/health, submit returns 201 with raw URL preserved, status moves
queued -> done in single-digit ms, all structured logs include
service / process_label / job_id / episode_id / stage with ms+Z
timestamps.

See CHANGELOG.md for the full review-pass findings list and fixes.

+++ backend/app/config.py
@@ -63,7 +63,7 @@ class Settings(BaseSettings):
 
     # Queue / HTTP / worker.
     QUEUE_POLL_INTERVAL_SECONDS: float = 2.0
-    JOB_TIMEOUT_SECONDS: int = 1800
+    JOB_TIMEOUT_SECONDS: float = 1800
     WEB_WORKERS: int = 2
 
     # Logging.
